chore: clean up connections protocol implementation in toolbox
