+++ src/renderer/components/AgentList.vue
@@ -72,6 +72,7 @@ import { new_connection } from '../connection_detail.js';
 import message_bus from '@/message_bus.js';
 import { base64_decode, base64_encode } from '../base64.js';
 import { from_store } from '../connection_detail.js';
+import 'ConnectionsProtocol.js';
 const uuidv4 = require('uuid/v4');
 const coordinate_mediation =
   (type) => `did:sov:BzCbsNYhMrjHiqZDTUASHg;spec/coordinate-mediation/1.0/${type}`;
@@ -299,95 +300,6 @@ export default {
         this.invitation_error = err.message;
       }
       this.new_agent_invitation = "";
-    },
-
-    async new_agent_invitation_process(invitation){
-      //process invite, prepare request
-      //extract c_i param
-      function getUrlVars(url) {
-        var vars = {};
-        var parts = url.replace(/[?&]+([^=&]+)=([^&]*)/gi, function(m,key,value) {
-          vars[key] = value;
-        });
-        return vars;
-      }
-
-      var invite_b64 = getUrlVars(invitation)["c_i"];
-      //base 64 decode
-      var invite_string = base64_decode(invite_b64);
-      var invite = JSON.parse(invite_string);
-
-      //make a did
-      const didcomm = new DIDComm.DIDComm();
-      await didcomm.Ready;
-      const toolbox_did = await didcomm.generateKeyPair();
-      toolbox_did.did = bs58.encode(Buffer.from(toolbox_did.publicKey.subarray(0, 16)));
-      toolbox_did.publicKey_b58 = bs58.encode(Buffer.from(toolbox_did.publicKey));
-      toolbox_did.privateKey_b58 = bs58.encode(Buffer.from(toolbox_did.privateKey));
-
-      let service_endpoint_block = {
-        "id": toolbox_did.did + ";indy",
-        "type": "IndyAgent",
-        "recipientKeys": [toolbox_did.publicKey_b58],
-        "serviceEndpoint": ""
-      };
-
-      if(this.mediator_connection && this.mediator_connection.active_as_mediator){
-        let mediator_agent = this.agent_list.find(a => a.active_as_mediator === true);
-        if(mediator_agent != null) {
-          service_endpoint_block["routingKeys"] = mediator_agent.mediator_info.routing_keys || [];
-          service_endpoint_block["serviceEndpoint"] = mediator_agent.mediator_info.endpoint;
-        }
-        console.log('Informing mediator about new key to route...');
-        await this.add_route_to_mediator(toolbox_did.publicKey_b58);
-      }
-
-      var req = {
-        "@id":  (uuidv4().toString()),
-        "~transport": {
-          "return_route": "all"
-        },
-        "@type": "did:sov:BzCbsNYhMrjHiqZDTUASHg;spec/connections/1.0/request",
-        "label": "ToolBox",
-        "connection": {
-          "DID": toolbox_did.did,
-          "DIDDoc": {
-            "@context": "https://w3id.org/did/v1",
-            "id": toolbox_did.did,
-            "publicKey": [{
-              "id": toolbox_did.did + "#keys-1",
-              "type": "Ed25519VerificationKey2018",
-              "controller": toolbox_did.did,
-              "publicKeyBase58": toolbox_did.publicKey_b58
-            }],
-            "service": [service_endpoint_block]
-          }
-        }
-      };
-      console.log("Exchange Request", req);
-
-      //send request, look for response
-      const packedMsg = await didcomm.packMessage(JSON.stringify(req), [bs58.decode(invite.recipientKeys[0])], toolbox_did, true);
-      console.log("Packed Exchange Request", packedMsg);
-
-      // this code assumes that the response comes via return-route on the post.
-      const res = await rp({
-        method: 'POST',
-        uri: invite.serviceEndpoint,
-        body: packedMsg,
-      });
-      const unpackedResponse = await didcomm.unpackMessage(res, toolbox_did);
-      const response = JSON.parse(unpackedResponse.message);
-      //TODO: Validate signature against invite.
-      let buff = new Buffer(response['connection~sig'].sig_data, 'base64');
-      let text = buff.toString('ascii');
-      //first 8 chars are a timestamp for the signature, so we ignore those before parsing value
-      response.connection = JSON.parse(text.substring(8));
-      console.log("response message", response);
-      let connection_detail = new_connection(invite.label, response.connection.DIDDoc, toolbox_did);
-      console.log("connection detail", connection_detail);
-      this.add_agent(connection_detail.to_store());
-      return connection_detail;
     },
   },
   created: async function(){
